chore(engine): removed plugins manager file, moved to engine since there was some duplicate code
