+++ engine/src/Options/Classes/Interactivity/Modes/Modes.ts
@@ -28,7 +28,7 @@ export class Modes implements IModes, IOptionLoader<IModes> {
             return;
         }
 
-        const interactors = this._engine.plugins.interactors.get(this._container);
+        const interactors = this._engine.interactors.get(this._container);
 
         if (!interactors) {
             return;
